refactor: CQRS 패턴 적용 및 서비스단 분리
